fix(commands): repair the option builders no test ever called
Counting every `pub fn -> Self` in the commands layer, not just the
`*Options` types, leaves 21 builders out of 362 that no test calls. Two
of them cannot work:

- `FtReducer::first_value_by` and `first_value_by_order` never emit the
  `BY` keyword. They send `FIRST_VALUE 2 @name @age` and
  `FIRST_VALUE 3 @name @age DESC`, both rejected with
  `Unknown argument @age at position 1 for FIRST_VALUE`. The counts and
  the keyword now come from a dedicated `FtFirstValue` clause.
- `RestoreOptions::frequency` emits `FREQUENCY 10.0` where `RESTORE`
  takes `FREQ 10`, so every call answers `syntax error`. The field is
  renamed and typed `u8`, the LFU counter's range.

`FT.AGGREGATE`'s `TOLIST` and `RANDOM_SAMPLE` reducers return an array
of strings per group, which `Vec<(String, String)>` cannot hold: both
failed with `CannotParseString`. `FtSearchResultRow::values` and
`extra_attributes` now hold an `FtAttributeValue`, either `Text` or
`Array`, read through `as_str()` / `as_array()`.

`FT.HYBRID`'s per-clause `YIELD_SCORE_AS` works on Redis 8.8 and is
covered on both the `SEARCH` and the `VSIM` clause.

All 362 builders are now called by a test.

+++ src/commands/generic_commands.rs
@@ -732,8 +732,8 @@ pub struct RestoreOptions {
     absttl: bool,
     #[serde(skip_serializing_if = "Option::is_none")]
     idletime: Option<i64>,
-    #[serde(skip_serializing_if = "Option::is_none")]
-    frequency: Option<f64>,
+    #[serde(rename = "FREQ", skip_serializing_if = "Option::is_none")]
+    frequency: Option<u8>,
 }
 
 impl RestoreOptions {
@@ -755,8 +755,10 @@ impl RestoreOptions {
         self
     }
 
+    /// Seeds the LFU access counter of the restored key, from 0 to 255.
+    /// Only meaningful under an LFU `maxmemory-policy`.
     #[must_use]
-    pub fn frequency(mut self, frequency: f64) -> Self {
+    pub fn frequency(mut self, frequency: u8) -> Self {
         self.frequency = Some(frequency);
         self
     }
